hw1/pyproject.toml
[tool.poetry]
name = "hw1"
version = "0.1.0"
description = ""
authors = ["Haohang Li <hli113@stevens.edu>"]
readme = "README.md"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "jorenham/jax_pep503"
url = "https://jorenham.github.io/jax_pep503/"
priority = "supplemental"

[tool.poetry.dependencies]
python = ">=3.10, < 3.11"
numpy = "^1.26.0"
pillow = "^10.0.1"
typer = "^0.9.0"
rich = "^13.6.0"
black = "^23.9.1"
jax = {extras = ["cpu"], version = "^0.4.16"}
ipykernel = "^6.25.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
